Prevent exception on truthiness of non-empty array
